Add readme to repo and add journal to .gitignore
in an effort to get this thing cleaned up and really presentable as a
repo, i'm going to start figuring out how to set this up in a way i'd be
proud of. i think i'm decided on learning git module and setting up the
projects as git modules to be consumed by the website, which will take
some reorganization, and i'm starting with my journal, which isn't
public facing yet anyway.